requirements.txt
asgiref==3.11.1
Django==5.2.13
django-tenants==3.8.0
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==25.1.0
packaging==26.1
Pillow==12.2.0
psycopg==3.2.13
psycopg-binary==3.2.13
PyJWT==2.12.1
python-dotenv==1.0.1
sqlparse==0.5.5
tzdata==2026.1
